auth-service-python/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
dnspython==2.4.2
gunicorn==21.2.0
